Avoid linting in `test` CI job.
We have a few different CI jobs setup:

* linting -> runs `yarn lint`
* test -> runs `yarn test`
* test-node -> runs `yarn test` for many Node versions

Because `yarn test` should **always** be the source of truth for "pass
vs fail" of a project, we have also setup `yarn test` to run `yarn
lint`. As you can see, this means that we are _actually_ running linting
in **all** of our test jobs now. 🥳

This updates the non-linting jobs to run `test:jest` directly (and avoid
running linting for every job).