Use @typescript-eslint/eslint-parser's recommended set of linting rules.
A few of these rules are slightly annoying (e.g. explicit return type),
but without using it's  `recommended` set we have maintanence issues
as we update eslint versions. We'd also have to manually manage the
specific built-in rules to disable (e.g. `no-unused-vars` in favor of
`@typescript-eslint/no-unused-vars` which is type aware).

I think overall, the balance is actually in the side of using
`@typescript-eslint/recommended`.