Ensure only open issues are used for avoiding to report a failure
Prior to this, we would find closed issues if they matched the other
criteria. This meant that a closed issue for the same `runId` would
prevent re-opening.

At the moment, since we include the `runId` directly in the issue title
this is "fine", however in the future we will likely be creating a
single issue and commenting / updating issue description for each
failure. In addition, this makes the tests _roughly_ impossible to write
(since it is impossible to reset to a state where recording the requests
would re-create an issue).