Run `eslint` directly for `lint-staged` configuration.
`lint-staged` passes the names of the staged files as positional
arguments to the end of the command specified in the configuration.
Since we were running `yarn lint` which always specifies `.` to be
linted (e.g. it calls `eslint .`) we were linting _all_ files and the
staged files twice (because it was running `eslint .
src/staged-file.ts`).

While I would prefer to avoid duplicating the specifie `eslint` options
that we pass I also prefer that `git commit` is "fast", and as this repo
grows (with the current setup) we would only get slower and slower.

These changes reduce the time to run `./node_modules/.bin/lint-staged`
(with a simple `.ts` file modification staged) from 2s to 1.5s
(avergaged over 10 runs).